move StyleCheckSettings to be used by several checks
